chore(test): prevent test compile OOM
